requirements.txt
psutil
pillow
ansible